move Help to component file
